# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indus Driving School API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingPackage =
  (typeof BookingPackage)[keyof typeof BookingPackage];

export const BookingPackage = {
  Morning: "Morning",
  Evening: "Evening",
} as const;

export interface Booking {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time slot (e.g. "7:00 AM") */
  time: string;
  package: BookingPackage;
  name: string;
  phone: string;
  createdAt: string;
}

export type CreateBookingRequestPackage =
  (typeof CreateBookingRequestPackage)[keyof typeof CreateBookingRequestPackage];

export const CreateBookingRequestPackage = {
  Morning: "Morning",
  Evening: "Evening",
} as const;

export interface CreateBookingRequest {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time slot (e.g. "7:00 AM") */
  time: string;
  package: CreateBookingRequestPackage;
  name: string;
  phone: string;
}

export interface ErrorResponse {
  error: string;
}