# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indus Driving School API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPackage } from "./bookingPackage";

export interface Booking {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time slot (e.g. "7:00 AM") */
  time: string;
  package: BookingPackage;
  name: string;
  phone: string;
  createdAt: string;
}