# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indus Driving School API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingRequestPackage } from "./createBookingRequestPackage";

export interface CreateBookingRequest {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time slot (e.g. "7:00 AM") */
  time: string;
  package: CreateBookingRequestPackage;
  name: string;
  phone: string;
}